Fix typo in overview of pitch notations
This text seems to be duplicated multiple times in the project, so this fix to just this file may not be enough. Searching for this section of the text and excluding the build folder, I think these are the other files that need to be changed:

* [`docs/ref/index.html` line 446](https://github.com/music-suite/music-suite.github.io/blob/master/docs/ref/index.html#L446)
* [`docs/ref/test.html` line 307](https://github.com/music-suite/music-suite.github.io/blob/master/docs/ref/test.html#L307)

I won’t change them in this pull request because I can’t do that from GitHub’s interface and I don’t want to bother downloading the repo to do it properly.